Refactoring: extracted new helper class to handle matching of instance method expectations on specific mocked instances.
