+++ main/src/mockit/internal/expectations/RecordAndReplayExecution.java
@@ -84,7 +84,7 @@ public final class RecordAndReplayExecution
             targetClasses.addAll(paramTypeRedefinitions.getTargetClasses());
          }
 
-         executionState.discoverMockedTypesToMatchOnInstances(targetClasses);
+         executionState.instanceBasedMatching.discoverMockedTypesToMatchOnInstances(targetClasses);
 
          if (dynamicPartialMocking != null && !dynamicPartialMocking.targetInstances.isEmpty()) {
             executionState.setDynamicMockInstancesToMatch(dynamicPartialMocking.targetInstances);
